tests: collect tests for all estimators in a single module
