+++ skchange/tests/test_all_detectors.py
@@ -1,12 +1,8 @@
 """Tests for all annotators/detectors in skchange."""
 
-from inspect import _empty, signature
-
 import numpy as np
 import pandas as pd
 import pytest
-from skbase.base import BaseObject
-from sktime.utils.estimator_checks import check_estimator, parametrize_with_checks
 
 from skchange.anomaly_detectors import ANOMALY_DETECTORS
 from skchange.base import BaseDetector
@@ -14,11 +10,6 @@ from skchange.change_detectors import CHANGE_DETECTORS
 from skchange.datasets.generate import generate_anomalous_data
 
 ALL_DETECTORS = ANOMALY_DETECTORS + CHANGE_DETECTORS
-
-
-@parametrize_with_checks(ALL_DETECTORS)
-def test_sktime_compatible_estimators(obj, test_name):
-    check_estimator(obj, tests_to_run=test_name, raise_exceptions=True)
 
 
 @pytest.mark.parametrize("Detector", ALL_DETECTORS)
@@ -97,22 +88,3 @@ def test_detector_not_implemented_methods():
         detector.dense_to_sparse(x)
     with pytest.raises(NotImplementedError):
         detector.sparse_to_dense(x, x.index, pd.Index(["a"]))
-
-
-@pytest.mark.parametrize("Detector", ALL_DETECTORS)
-def test_detector_no_mutable_defaults(Detector: BaseDetector):
-    """Ensure no detectors have mutable default arguments."""
-
-    detector = Detector.create_test_instance()
-    sig = signature(detector.__init__)
-    mutable_types = (
-        list,
-        dict,
-        set,
-        BaseObject,
-    )
-    for param in sig.parameters.values():
-        if param.default is not _empty and isinstance(param.default, mutable_types):
-            raise AssertionError(
-                f"Mutable default argument found in {Detector.__name__}: {param.name}"
-            )
