+++ skchange/tests/test_all_interval_evaluators.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pandas as pd
 import pytest
-from sktime.utils.estimator_checks import check_estimator, parametrize_with_checks
 
 from skchange.anomaly_scores import ANOMALY_SCORES
 from skchange.change_scores import CHANGE_SCORES
@@ -9,11 +8,6 @@ from skchange.costs import COSTS
 from skchange.datasets import generate_alternating_data, generate_anomalous_data
 
 INTERVAL_EVALUATORS = COSTS + CHANGE_SCORES + ANOMALY_SCORES
-
-
-@parametrize_with_checks(INTERVAL_EVALUATORS)
-def test_sktime_compatible_estimators(obj, test_name):
-    check_estimator(obj, tests_to_run=test_name, raise_exceptions=True)
 
 
 @pytest.mark.parametrize("Evaluator", INTERVAL_EVALUATORS)
